docshell/test/iframesandbox/test_our_auxiliary_navigation_by_location.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=785310
html5 sandboxed iframe should not be able to perform top navigation with scripts allowed
-->
<head>
<meta charset="utf-8">
<title>Test for Bug 785310 - iframe sandbox our auxiliary navigation by location tests</title>
<script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

<script>
  SimpleTest.waitForExplicitFinish();

  function runScriptNavigationTest(testCase) {
    window.onmessage = function(event) {
      if (event.data != "ourWindow") {
        ok(false, "event.data: got '" + event.data + "', expected 'ourWindow'");
      }
      ok(!testCase.shouldBeBlocked, testCase.desc, "auxiliary navigation was NOT blocked");
      runNextTest();
    };
    try {
      SpecialPowers.wrap(window["testIframe"]).eval(testCase.script);
    } catch(e) {
      ok(testCase.shouldBeBlocked, testCase.desc, SpecialPowers.wrap(e).message);
      runNextTest();
    }
  }

  var testCaseIndex = -1;
  testCases = [
    {
      desc: "Test 1: location.replace on auxiliary opened by us should NOT be blocked",
      script: "openedWindow.location.replace('file_our_auxiliary_navigation_by_location.html')",
      shouldBeBlocked: false
    },
    {
      desc: "Test 2: location.assign on auxiliary opened by us should be blocked without allow-same-origin",
      script: "openedWindow.location.assign('file_our_auxiliary_navigation_by_location.html')",
      shouldBeBlocked: true
    },
    {
      desc: "Test 3: location.href on auxiliary opened by us should NOT be blocked",
      script: "openedWindow.location.href = 'file_our_auxiliary_navigation_by_location.html'",
      shouldBeBlocked: false
    },
    {
      desc: "Test 4: location.hash on auxiliary opened by us should be blocked without allow-same-origin",
      script: "openedWindow.location.hash = 'wibble'",
      shouldBeBlocked: true
    }
  ];

  function runNextTest() {
    ++testCaseIndex;
    if (testCaseIndex == testCases.length) {
      SpecialPowers.wrap(window["testIframe"]).eval("openedWindow.close()");
      SimpleTest.finish();
      return;
    }

    runScriptNavigationTest(testCases[testCaseIndex]);
  }

  window.onmessage = runNextTest;

  window.onload = function() {
    SpecialPowers.wrap(window["testIframe"]).eval("var openedWindow = window.open('file_our_auxiliary_navigation_by_location.html', 'ourWindow')");
  }
</script>

</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=785310">Mozilla Bug 785310</a>
<p id="display"></p>
<div id="content">
Tests for Bug 785310
</div>

<iframe name="testIframe" sandbox="allow-scripts allow-popups"></iframe>
</body>
</html>
